Introducing Wifi Reconnect

+++ parameters.h
@@ -134,7 +134,7 @@ struct strConfig {
   void WriteStringToEEPROM(int beginaddress, String string){
     char  charBuf[string.length() + 1];
     string.toCharArray(charBuf, string.length() + 1);
-    for (int t =  0; t < sizeof(charBuf); t++)
+    for (uint16_t t =  0; t < sizeof(charBuf); t++)
     {
       EEPROM.write(beginaddress + t, charBuf[t]);
     }
